daily; table adjustment

+++ models/posts.cs
@@ -1,4 +1,5 @@
 using System.ComponentModel.DataAnnotations;
+using System.Numerics;
 
 namespace vibe_backend.models 
 {
@@ -6,10 +7,10 @@ namespace vibe_backend.models
     {
         [Key]
         public int post_table_id {get; set;}
-        public int user_id { get; set; }
+        public BigInteger user_id { get; set; }
         public string content { get; set; }
 
-        public int post_id { get; set; }
+        public BigInteger post_id { get; set; }
 
         public DateTime post_date { get; set; }
     }
